Update 092316 2
Version bump to 1.5.6 for repo push
- Added 3DO
- Added Atari ST
- Added Dinothawr
- Updated External Launch Database
- Updated Wizard
- Updated function that searches for local files for launching, this
fixes quite a few issues with launching local files instead of
downloading.  Should help users using IARL just as a front-end for
games they have on their local computer
- Added functionality to use games hosted from Google Drive or Dropbox
- Added an advanced function to use python subprocess to provide
external commands.  This is by default not used currently, but could
help with troubleshooting for users in the future.